fixed project structure
