+++ src/App.js
@@ -1,15 +1,15 @@
 import {BrowserRouter as Router,Routes,Route,Link} from "react-router-dom";
-import Login from './components/scripts/login';
-import Dashboard from './components/scripts/dashboard';
-import Survey from './components/scripts/survey';
-import Register from './components/scripts/Register';
-import Search from './components/scripts/search';
-import FormComponent from './FormContributor.js';
+import Login from './components/Login';
+import Dashboard from './components/Dashboard';
+import Survey from './components/Survey';
+import Register from './components/Register';
+import Search from './components/Search';
+import FormComponent from './components/FormContributor.js';
 import Insights from "./components/scripts/insights";
 import Profile from "./components/Profile";
-import ContributorPage from "./components/scripts/ContributorFeed";
-import LikedArticles from "./components/scripts/LikedArticles";
-import { OpenAI } from './openAI'
+import ContributorPage from "./components/ContributorFeed";
+import LikedArticles from "./components/LikedArticles";
+import { OpenAI } from './components/scripts/openAI'
 
 function App() {
   return (
