+++ src/components/scripts/openAI.js
